Commiting for the work, but ditching the idea
The proxy has been way to problematic. In order for it to work I need to
cover almsot any and all protocol that will be redirected through my
program. Where this would be the preferred method. I don't have the time
for me to try to get this to work in 10 weeks. The resources on this are
very small and I can't find much with helping me on this. So I am going
back to my original idea, hopefully this doesn't give me as many
problems.