A lot of resources added to Todo.txt
I have found a lot of different resources on trying to log out a
specific user. I feel like I am getting closer to at least a way of
doing it by reinventing the wheel. I can't find any other resources that
provide the functionality that I need. It is realitivly simple yet I
cant find anything on it. So I am looking into WMI to see if I can find
the logged on Session ID. So I can actually use my UserLogoff method.
That is all I need yet I can't find any information on it or where to
get it. I could probably get it from the cmd, but I would have to parse
strings in order to get that information out, which would be way harder
than to call something like the WMI. So Hopefully it is in there.