Week 2 Completed
Now that I got everything to work for week 2 there are a few things that
I want to get done to make sure that it will be ready for web. For one I
will need to get the database to work. I will also need to get the
events for logging in to work so that way I can listen to windows and
start my program. I will also need to make it a service, but that will
be easy I beleive.