Rest works
Need to form an API system but as of yet I can go to
localhost:3000/computers.json with the payload of auth_token
5dxGZpsvMqjHwHXn4W8n and get the computers from the db. :D

However in doing this I have somehow broken the web login. I am looking
into this as the interweb says that it can be fixed via a method in the
application controller.

This needs to be fixed before the deadline otherwise it will be hard to
show something has happened.