Limitations on Json added
These limitations for sending sensitive data has been added. Also There
seems to be a problem with my account_processes. In my next commit I
will need to recreate it and see if that fixes it. Rails says there
isn't a class by the name of Account Proces <- only one S. I am not sure
why it is doing that. I have gone through all my code and I have not
been able to find anywhere I set it with only 1 s. So I will just
recreate it. I know there is data in the db for it so It isn't like it
is throwing nulls.

Besides that, my rails API is almost complete. I will just need to setup
the desktop portion to parse and store the data. :D which is going to be
a pain.