Updated JSON Attributes to lowercase names
Added these attributes to lowercase the name of the properties. This was
causing an issue with sending it to my api. This has been fixed.
(Possibly an issue?)

Also added an attribute to Lists<object> in my model classes. I added
NullValueHandling, this was because a user might not always specify a
list of computers for an account, but they would like to update their
account information. This makes it easier and I am sure glad I found it
after 2 hours of work and reasearch.