Everything is broken, Saving whats been done
I am going to create a new project that manages the REST portion of the
project. This will create some sepparation between the web service and
my desktop service. I believe this is the best solution to my problem.
There are certain attributes that json needs me to add in order for it
to work. This should simplify my life once I get it to work.