RavenDb was added Removed Sqlite
I have given up on Sqlite, it is giving (once again) way too many
problems with compatibility. Entity Framework is supposed to work with
it, but the lack of ducumentation and with my situation I cant figure
out how to fix it. I have found an alternative solution which works for
desktop application and is also native with JSON. This is retarded.