The next set of commits involved refactoring.
For right now I am going to be pushing refactoring commits. I am going
to be simplifying my application so I can get it ready for capstone
presentations. This includes removing Programs model and moving history
class to associate with a computer.

History is being associated with computer because the history is global.
If i were able to use a proxy I would be able to associate an account.

For Program, It doesn't really serve a purpose. I don't have a sure way
of identifying if it is an actually application or just a process with a
title. So It is being removed for Process replacement.