Saving all this crap.
Gui doesn't work in windows XP as for why? Not sure. I tried running it
on my desktop which has windows 8 but no development system on it. It
didn't run the gui. So I guess I will just demo my project with out one.
I figured i would work on the gui so I could work on something else that
might work. And I could start to tie things together. But ugh