Added Timers to account and made  checks on account
Still a lot of work needing to be done. However I now have it so that it
has a few timers. Some of the problems I am running accross are DB multi
threaded issues.

Anyways, still a work in progress, Added a Locked variable to account so
that I know if it is locked. dsafdsagfg Lots to do.