Update on time fixed
I have fixed the Service to check from the db on if the account exists.
This can be very problematic. Since my db isn't thread safe nore is
there much informtion on getting it to that point. It could
potentionally explode on conflicts. That might actually be why I have to
restart it multiple times before I get it to work.