Not Pretty in the least
Well, I have somewhat fixed some things. The db keeps throwing errors
that I don't know how to fix. The stuff I find on google says that it
shouldn't be doing that in the first place. Also if the gui becomes the
main project, there are certain bugs that are triggered. I am thinking
it might have to do with threading. However I only have 1 thread.