Committing some refactoring
So there were all those problems in early commits.
I am going through my code and refactoring my methods to better handle
these errors. Some of which I am just try catching and hoping that I
don't lose any data. Over the time of working on this for the rest of
the week I will be monitoring this and making adjustments.